my_doccano2/my_doccano2/app/requirements.txt
absl-py==0.8.1
amqp==1.4.9
anyjson==0.3.3
astor==0.8.0
billiard==3.3.0.23
boto==2.49.0
boto3==1.10.28
botocore==1.13.28
celery==3.1.18
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.1
django-cors-headers==3.1.1
django-filter==2.2.0
django-redis==4.10.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
docutils==0.15.2
gast==0.3.2
gensim==3.8.0
grpcio==1.25.0
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
kombu==3.0.37
Markdown==3.1.1
MarkupSafe==1.1.1
more-itertools==7.2.0
mysqlclient==1.4.4
numpy==1.14.5
Pillow==6.2.1
protobuf==3.11.0
PyJWT==1.7.1
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.3
redis==2.10.6
requests==2.22.0
s3transfer==0.2.1
scipy==1.3.3
six==1.13.0
smart-open==1.9.0
tensorboard==1.10.0
tensorflow==1.10.0`
termcolor==1.1.0
uritemplate==3.0.0
urllib3==1.25.6
vine==1.3.0
Werkzeug==0.16.0
zipp==0.6.0
